src: Made code less repetitive.
