+++ .eslintrc.js
@@ -132,7 +132,7 @@ module.exports = {
   overrides: [
     {
       // Using the func module has its costs, but it is worth it.
-      files: ["./src/cli/commands/*.ts", "./src/cli/options/*.ts"],
+      files: ["./src/cli/commands/**/*.ts", "./src/cli/options/**/*.ts"],
       rules: {
         "@typescript-eslint/no-extraneous-class": "off",
         "no-console": "off",
