+++ src/cli/commands/Help.ts
@@ -1,0 +1,8 @@
+import { Command } from "func";
+
+import BaseHelp from "./base/BaseHelp";
+
+@Command({ name: "help", alias: "h" })
+class Help extends BaseHelp { }
+
+export default Help;
